\name{getChrInfo}
\alias{getChrInfo}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
get the chromosome location information of methylation probes
}
\description{
get the chromosome location information of methylation probes
}
\usage{
getChrInfo(methyData, lib = NULL, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{methyData}{
a MethyLumiM object
}
  \item{lib}{
Methylation annotation library
}
\item{\dots}{ optional arguments to \code{\link{addAnnotationInfo}}.}
}
\value{
a data.frame  
}
\author{
Pan Du
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{methods}
