\name{lumi-package}
\alias{lumi-package}
\alias{lumi}
\docType{package}
\title{
A package for preprocessing Illumina microarray data
}
\description{
lumi R package is designed to preprocess the Illumina microarray (BeadArray) data. It includes functions of Illumina data input, quality control, variance stabilization, normalization and gene annotation. 
}
\details{
	\tabular{ll}{
		Package: \tab lumi\cr
		Type: \tab Package\cr
		Version: \tab 1.1.0\cr
		Date: \tab 2007-03-23\cr
		License: \tab  LGPL version 2 or newer \cr
	}

}
\author{
Pan Du, Simon Lin
Maintainer: Pan Du <dupan@northwestern.edu>
}
\references{
1. Du, P., Kibbe, W.A. and Lin, S.M., (2008) 'lumi: a pipeline for processing Illumina microarray', Bioinformatics 24(13):1547-1548

2. Lin, S.M., Du, P., Kibbe, W.A., (2008) 'Model-based  Variance-stabilizing Transformation for Illumina Microarray Data', Nucleic Acids Res. 36, e11

3. Du, P., Kibbe, W.A. and Lin, S.M., (2007) 'nuID: A universal naming schema of oligonucleotides for Illumina, Affymetrix, and other microarrays', Biology Direct, 2, 16

}
\keyword{ package }

\seealso{

}
\examples{

}
