\name{MethyLumiM2GenoSet}
\alias{MethyLumiM2GenoSet}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Coerce objects of MethyLumiM-class to MethyGenoSet
}
\description{
    Coerce objects of \code{\link{MethyLumiM-class}} to \code{MethyGenoSet}
}
\usage{
MethyLumiM2GenoSet(methyLumiM, lib = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{methyLumiM}{
a MethyLumiM object
}
  \item{lib}{
lib is a annotation library
}
}
\details{

}
\value{
a MethyGenoSet object
}
\references{

}
\author{
Pan Du
}

\seealso{
 \code{\link{MethyGenoSet}}
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{methods}
