\name{lumi-package}
\alias{lumi-package}
\alias{lumi}
\docType{package}
\title{
A package for preprocessing Illumina microarray data
}
\description{
lumi R package is designed to preprocess the Illumina microarray (BeadArray) data. It includes functions of Illumina data input, quality control, variance stabilization, normalization and gene annotation. 
}
\details{
	\tabular{ll}{
		Package: \tab lumi\cr
		Type: \tab Package\cr
		Version: \tab 1.0.2\cr
		Date: \tab 2007-01-01\cr
		License: \tab  LGPL version 2 or newer \cr
	}

}
\author{
Pan Du, Simon Lin
Maintainer: Pan Du <dupan@northwestern.edu>
}
\references{
1. Lin, S.M., Du, P., Kibbe, W.A.,  "Model-based Variance-stabilizing Transformation for Illumina Mi-croarray Data", submitted

2. Du, P., Kibbe, W.A. and Lin, S.M., "nuID: A Novel Identifier for Oligos, Ideal for Oligonucleotide-based  Microarrays", submitted.

}
\keyword{ package }

\seealso{

}
\examples{

}
