\name{m2beta}
\alias{m2beta}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Convert methylation M-value to Beta-value
}
\description{
Convert methylation M-value to Beta-value
}
\usage{
m2beta(m)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{m}{
	a matrix or vector of methylation M-value
}
}
\details{
Convert methylation M-value to Beta-value
}
\value{
return methylation Beta-value with the same size of input M-value
}
\references{

}
\author{
Pan Du
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
See Also as \code{\link{beta2m}}
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }

